moved PoolPairs.tsx to components/dex because it's not common at all
