+++ src/components/dex/PoolPairs.tsx
@@ -6,7 +6,7 @@ import { getAssetIcon } from '@components/icons/assets'
 import NumberFormat from 'react-number-format'
 import BigNumber from 'bignumber.js'
 
-export function PoolPairsTable ({ poolPairs }: {poolPairs: PoolPairData[]}): JSX.Element {
+export function PoolPairsTable ({ poolPairs }: { poolPairs: PoolPairData[] }): JSX.Element {
   return (
     <AdaptiveTable className='mt-6'>
       <AdaptiveTable.Header>
@@ -17,9 +17,13 @@ export function PoolPairsTable ({ poolPairs }: {poolPairs: PoolPairData[]}): JSX
         <AdaptiveTable.Head>
           <div className='flex items-center justify-end'>
             <div>APR</div>
-            <HoverPopover popover='On defiscan.live, only block rewards are included in the APR calculation. With commission, the expected APR is much higher. We will update this soon.'>
+            <HoverPopover
+              popover='On defiscan.live, only block rewards are included in the APR calculation. With commission, the expected APR is much higher. We will update this soon.'
+            >
               <div className='p-1 cursor-help'>
-                <IoAlertCircle className='h-4 w-4 text-black opacity-60 group-hover:text-primary group-hover:opacity-100' />
+                <IoAlertCircle
+                  className='h-4 w-4 text-black opacity-60 group-hover:text-primary group-hover:opacity-100'
+                />
               </div>
             </HoverPopover>
           </div>
